ElectricityDemandForecasting/data-download/requirements.txt
pandas
matplotlib
requests
python-dateutil
python-dotenv
openpyxl
certifi
charset-normalizer
idna
six
urllib3
